(chore) use Parcel to build the app (#38)
Because where's the fun in building a webapp if you're not going to swap
out tooling all the time?

Swaps out the old gulp-based workflow for Parcel, which covers pretty
much everything we need with zero configuration. But this should help us
with more complex things in the future like adding in GOAT templating
(#2) and better performance by serving ES modules to browsers that
support it.

Note that with this commit alone, the bundle size is **worse** than it
used to be. Parcel's image optimization is lossless by default and
doesn't minify the massive 4.4MB `tyler-paulo.jpg` as much as the
imagemin-based approach (#15). This will be fixed up in a future commit.

Also there's a slight dev workflow change here, but overall should be
simpler:
- `npm start` to run the dev server
- `npm run build` to build for production

Co-authored-by: YellowKirby <YellowKirby@users.noreply.github.com>